called the components of react
Updated Math magic project
