+++ src/App.css
@@ -1,0 +1,5 @@
+.App {
+  display: grid;
+  place-items: center;
+  height: 95vh;
+}
