+++ src/App.js
@@ -1,0 +1,12 @@
+import Calculator from './components/Calculator';
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <Calculator />
+    </div>
+  );
+}
+
+export default App;
