+++ src/index.css
@@ -1,0 +1,34 @@
+html {
+  box-sizing: border-box;
+  font-size: 100%;
+}
+
+*,
+*::before,
+*::after {
+  box-sizing: inherit;
+}
+
+body {
+  --color-orange: tomato;
+  --color-dark: rgb(24, 24, 24);
+  --color-dark-grey: rgb(81, 81, 81);
+  --color-grey: rgb(130, 130, 130);
+  --color-light-grey: rgb(237, 237, 237);
+  --color-light: rgb(246, 246, 246);
+  --color-white: white;
+
+  margin: 0;
+  padding: 0;
+  font-size: 0.85em;
+  font-family: sans-serif;
+  font-weight: 400;
+  color: var(--color-dark-grey);
+  background-color: var(--color-light);
+}
+
+@media (min-width: 64em) {
+  body {
+    font-size: 1em;
+  }
+}
